added simple query, Dto is not used yet

+++ src/main/java/am/hakobyan/hgh/repository/TaskRepository.java
@@ -3,5 +3,5 @@ package am.hakobyan.hgh.repository;
 import am.hakobyan.hgh.entity.Task;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TaskRepository extends CrudRepository<Task, Integer> {
+public interface TaskRepository extends CrudRepository<Task, Long> {
 }
